pagination and delete files fix

+++ app/jobs/delete_jpg_files_job.rb
@@ -4,8 +4,8 @@ class DeleteJpgFilesJob < ApplicationJob
   def perform(files)
     files.each do |file|
       basename = file.file.file.basename
-      directory = Rails.root.join("tmp", "jpgs", basename)
-      FileUtils.rm_rf(directory)
+      directory = Rails.root.join("tmp", "jpgs", basename[0...-7])
+      FileUtils.rm_rf(directory) if directory.present?
     end
   end
 end
